fix: Move `rate_limiter` into module under `vmm`
