--- src/rate_limiter/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "rate_limiter"
-version = "0.1.0"
-authors = ["Amazon Firecracker team <firecracker-devel@amazon.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[dependencies]
-timerfd = "1.2.0"
-versionize = "0.1.6"
-versionize_derive = "0.1.4"
-
-logger = { path = "../logger" }
-snapshot = { path = "../snapshot" }
-utils = { path = "../utils" }
