+++ src/vmm/Cargo.toml
@@ -27,7 +27,6 @@ timerfd = "1.2.0"
 
 logger = { path = "../logger" }
 mmds = { path = "../mmds" }
-rate_limiter = { path = "../rate_limiter" }
 seccompiler = { path = "../seccompiler" }
 snapshot = { path = "../snapshot"}
 utils = { path = "../utils" }
