+++ src/vmm/src/devices/virtio/block/test_utils.rs
@@ -8,7 +8,7 @@ use std::thread;
 #[cfg(test)]
 use std::time::Duration;
 
-use rate_limiter::RateLimiter;
+use crate::rate_limiter::RateLimiter;
 use utils::kernel_version::{min_kernel_version_for_io_uring, KernelVersion};
 use utils::tempfile::TempFile;
 use vm_memory::{Bytes, GuestAddress};
