+++ src/vmm/src/devices/virtio/net/persist.rs
@@ -11,8 +11,8 @@ use logger::warn;
 use mmds::data_store::Mmds;
 use mmds::ns::MmdsNetworkStack;
 use mmds::persist::MmdsNetworkStackState;
-use rate_limiter::persist::RateLimiterState;
-use rate_limiter::RateLimiter;
+use crate::rate_limiter::persist::RateLimiterState;
+use crate::rate_limiter::RateLimiter;
 use snapshot::Persist;
 use utils::net::mac::{MacAddr, MAC_ADDR_LEN};
 use versionize::{VersionMap, Versionize, VersionizeResult};
