+++ src/vmm/src/rpc_interface.rs
@@ -1111,8 +1111,8 @@ mod tests {
         pub fn update_block_rate_limiter(
             &mut self,
             _: &str,
-            _: rate_limiter::BucketUpdate,
-            _: rate_limiter::BucketUpdate,
+            _: crate::rate_limiter::BucketUpdate,
+            _: crate::rate_limiter::BucketUpdate,
         ) -> Result<(), VmmError> {
             Ok(())
         }
@@ -1120,10 +1120,10 @@ mod tests {
         pub fn update_net_rate_limiters(
             &mut self,
             _: &str,
-            _: rate_limiter::BucketUpdate,
-            _: rate_limiter::BucketUpdate,
-            _: rate_limiter::BucketUpdate,
-            _: rate_limiter::BucketUpdate,
+            _: crate::rate_limiter::BucketUpdate,
+            _: crate::rate_limiter::BucketUpdate,
+            _: crate::rate_limiter::BucketUpdate,
+            _: crate::rate_limiter::BucketUpdate,
         ) -> Result<(), VmmError> {
             if self.force_errors {
                 return Err(VmmError::DeviceManager(
